Add auth

+++ src/main/scala/com/anilsenay/controllers/BrandController.scala
@@ -1,17 +1,15 @@
 package com.anilsenay.controllers
 
-import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import com.anilsenay.models.Brand
 import com.anilsenay.services.BrandService
-import com.typesafe.scalalogging.LazyLogging
 import spray.json._
 
 import scala.util.{Failure, Success}
 
-class BrandController(dbService: BrandService.type) extends SprayJsonSupport with DefaultJsonProtocol with LazyLogging {
+class BrandController(dbService: BrandService.type) extends BaseController {
   val route: Route = pathPrefix("api" / "brand") {
     get {
       pathEndOrSingleSlash {
@@ -20,17 +18,22 @@ class BrandController(dbService: BrandService.type) extends SprayJsonSupport wit
     } ~
     post {
       pathEndOrSingleSlash {
-      entity(as[Brand]) { brand =>
-        val saved = dbService.insertBrand(brand.brandName)
-        onComplete(saved) {
-          case Success(updatedRows) => {
-            logger.info(s"Inserted brand")
-            complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
-          }
-          case Failure(e) => {
-            logger.error(s"Failed to insert brand", e)
-            complete(StatusCodes.InternalServerError)
+        entity(as[Brand]) { brand =>
+          authenticate { authUser =>
+            if (authUser.isAdmin) {
+              val saved = dbService.insertBrand(brand.brandName)
+              onComplete(saved) {
+                case Success(updatedRows) => {
+                  logger.info(s"Inserted brand")
+                  complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
+                }
+                case Failure(e) => {
+                  logger.error(s"Failed to insert brand", e)
+                  complete(StatusCodes.InternalServerError)
+                }
+              }
             }
+            else complete(StatusCodes.Unauthorized)
           }
         }
       }
@@ -38,26 +41,36 @@ class BrandController(dbService: BrandService.type) extends SprayJsonSupport wit
     put {
       path(LongNumber) { id =>
         entity(as[Brand]) { brand =>
-          val updated = dbService.update(id, brand.brandName)
-          onComplete(updated) {
-            case Success(updatedRows) => complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
-            case Failure(e) => {
-              logger.error(s"Failed to update brand ${id}", e)
-              complete(StatusCodes.InternalServerError)
+          authenticate { authUser =>
+            if (authUser.isAdmin) {
+              val updated = dbService.update(id, brand.brandName)
+              onComplete(updated) {
+                case Success(updatedRows) => complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
+                case Failure(e) => {
+                  logger.error(s"Failed to update brand ${id}", e)
+                  complete(StatusCodes.InternalServerError)
+                }
+              }
             }
+            else complete(StatusCodes.Unauthorized)
           }
         }
       }
     } ~
     delete {
       path(LongNumber) { id =>
-        val deleted = dbService.delete(id)
-        onComplete(deleted) {
-          case Success(updatedRows) => complete(JsObject("deletedRows" -> JsNumber(updatedRows)))
-          case Failure(e) => {
-            logger.error(s"Failed to delete brand ${id}", e)
-            complete(StatusCodes.InternalServerError)
+        authenticate { authUser =>
+          if (authUser.isAdmin) {
+            val deleted = dbService.delete(id)
+            onComplete(deleted) {
+              case Success(updatedRows) => complete(JsObject("deletedRows" -> JsNumber(updatedRows)))
+              case Failure(e) => {
+                logger.error(s"Failed to delete brand ${id}", e)
+                complete(StatusCodes.InternalServerError)
+              }
+            }
           }
+          else complete(StatusCodes.Unauthorized)
         }
       }
     }
